@@ -14,7 +14,7 @@ description: Change the world.
 version: 1.3.0-rc.0+21
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
