@@ -10,12 +10,11 @@ class FileSystemGridView extends StatelessWidget {
   final AssetOpenedCallback onOpened;
   final VoidCallback onRefreshed;
   const FileSystemGridView(
-      {Key? key,
+      {super.key,
       required this.assets,
       required this.selectedPath,
       required this.onOpened,
-      required this.onRefreshed})
-      : super(key: key);
+      required this.onRefreshed});
 
   @override
   Widget build(BuildContext context) {
